requirements.txt
tornado==6.0.1
python-jose==3.0.1
requests==2.21.0
pytest_runner>=4.2
pytest==3.6.1
pytest-xdist==1.23.2
pytest-cov==2.6.0
teamcity-messages==1.21